feat: body and query vlidation refactored to middleware/Validation
