+++ src/server/controllers/cities/Create.ts
@@ -1,6 +1,5 @@
 import { Request, RequestHandler, Response } from "express";
-import { StatusCodes } from "http-status-codes";
-// import { validation } from "../../shared/middleware";
+import { validation } from "../../shared/middleware";
 import * as yup from "yup";
 
 interface ICities {
@@ -13,24 +12,6 @@ const bodyValidation: yup.Schema<ICities> = yup.object().shape({
   state: yup.string().required().min(2),
 });
 
-// express middleware
-export const createBodyValidator: RequestHandler = async (req, res, next) => {
-  try {
-    await bodyValidation.validate(req.body, { abortEarly: false });
-    return next();
-  } catch (err) {
-    const yupError = err as yup.ValidationError;
-    const validationErrors: Record<string, string> = {};//
-
-    yupError.inner.forEach(error => {
-      if (error.path === undefined) return;
-      validationErrors[error.path] = error.message;
-    });
-
-    return res.status(StatusCodes.BAD_REQUEST).json({ errors: validationErrors });
-  }
-};
-
 interface IFilter {
   filter?: string;
 }
@@ -38,26 +19,9 @@ interface IFilter {
 const queryValidation: yup.Schema<IFilter> = yup.object().shape({
   filter: yup.string().required().min(3)
 });
-export const createQueryValidator: RequestHandler = async (req, res, next) => {
-  try {
-    await queryValidation.validate(req.quey, { abortEarly: false });
-    return next();
-  } catch (err) {
-    const yupError = err as yup.ValidationError;
-    const validationErrors: Record<string, string> = {};//
 
-    yupError.inner.forEach(error => {
-      if (error.path === undefined) return;
-      validationErrors[error.path] = error.message;
-    });
-
-    return res.status(StatusCodes.BAD_REQUEST).json({ errors: validationErrors });
-  }
-};
-
-
-
-// export const createValidation = validation();
+export const createValidation = validation(queryValidation);
+export const createBodyValidation = validation(bodyValidation);
 
 // controller (only executed after bodyValidation)
 export const create: RequestHandler = async (req: Request<{}, {}, ICities>, res: Response) => {
