+++ src/server/shared/middleware/index.ts
@@ -1,1 +1,1 @@
-// export * from "./Validation";
+export * from "./Validation";
